+++ .ondevice/storybook.requires.js
@@ -1,9 +1,10 @@
 /* do not change this file, it is auto generated by storybook. */
 
-import { configure } from '@storybook/react-native';
+import { configure, addDecorator, addParameters, addArgsEnhancer } from '@storybook/react-native';
 
 const getStories = () => {
 	return [
+		require('../app/containers/Avatar/Avatar.stories.js'),
 		// require("../app/containers/BackgroundContainer/index.stories.js"),
 		require('../app/containers/Button/Button.stories.js')
 		// require("../app/containers/LoginServices/LoginServices.stories.js"),
