+++ storybook/stories/index.js
@@ -12,7 +12,7 @@ import './HeaderButtons';
 import './UnreadBadge';
 import '../../app/views/ThreadMessagesView/Item.stories.js';
 import '../../app/views/DiscussionsView/Item.stories.js';
-import './Avatar';
+// import './Avatar.stories';
 import './NewMarkdown';
 import '../../app/containers/BackgroundContainer/index.stories.js';
 import '../../app/containers/RoomHeader/RoomHeader.stories.js';
